Move line and arrow editors to new mod
